+++ man/examples/findRelated.R
@@ -1,0 +1,5 @@
+# Load example data
+bg <- BGData:::loadExample()
+
+G <- getG(bg@geno)
+findRelated(G)